requirements.txt
asgiref
beautifulsoup4
certifi
charset-normalizer
Django
django-cors-headers
django-environ
djangorestframework
djangorestframework-simplejwt
docopt
# GDAL @ file:///C:/Users/user/pipwin/GDAL-3.4.3-cp311-cp311-win_amd64.whl#sha256=f78861fb5115d5c2f8cf3c52a492ff548da9e1256dc84088947379f90e77e5b6
idna
Js2Py
packaging
Pillow
pipwin
psycopg2
pyjsparser
PyJWT
PyPrind
pySmartDL
pytz
requests
six
soupsieve
sqlparse
tzdata
tzlocal
urllib3
